backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-cloud-aiplatform
google-generativeai
google-api-core>=2.27.0,<3.0.0
